+++ DKBasicEngine 1.0/Core/Components/Camera.cs
@@ -2,6 +2,7 @@
 * (C) 2017 David Knieradl 
 */
 
+using DKBasicEngine_1_0.Core.Ext;
 using System;
 using System.Collections.Generic;
 using System.Drawing;
@@ -10,9 +11,9 @@ using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace DKBasicEngine_1_0
+namespace DKBasicEngine_1_0.Core.Components
 {
-    public class Camera
+    public sealed class Camera
     {
         public Vector3 Position;
