+++ DKBasicEngine 1.0/Core/Scene.cs
@@ -3,20 +3,15 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 
-namespace DKBasicEngine_1_0
+namespace DKBasicEngine_1_0.Core
 {
     public class Scene : IPage
     {
         public string Name = "";
 
-        private Stopwatch TimeOutControls = new Stopwatch();
-        private TimeSpan TimeOut = new TimeSpan(0, 0, 1);
-        
-        internal readonly List<GameObject> Model            = new List<GameObject>();
-        internal readonly List<GameObject> NewlyGenerated   = new List<GameObject>();
-
-        /*public int FocusSelection { get; set; } = 0;
-        public List<IControl> PageControls { get; } = new List<IControl>();*/
+        internal readonly List<GameObject> Model          = new List<GameObject>();
+        internal readonly List<GameObject> NewlyGenerated = new List<GameObject>();
+        //internal readonly List<GameObject> AllGameObjects = new List<GameObject>();
 
         public Scene()
         { }
