+++ DKBasicEngine 1.0/Core/Script.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using DKBasicEngine_1_0.Core.Components;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DKBasicEngine_1_0
+namespace DKBasicEngine_1_0.Core
 {
     public abstract class Script
     {
