+++ DKBasicEngine 1.0/Core/Scripts/ButtonScript.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using DKBasicEngine_1_0.Core.UI;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DKBasicEngine_1_0
+namespace DKBasicEngine_1_0.Core.Scripts
 {
     internal sealed class ButtonScript : Script
     {
