+++ DKBasicEngine 1.0/Core/SystemExt/Extensions.cs
@@ -2,12 +2,13 @@
 * (C) 2017 David Knieradl 
 */
 
+using DKBasicEngine_1_0.Core;
 using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace DKBasicEngine_1_0
+namespace DKBasicEngine_1_0.Core.Ext
 {
     public static class Extensions
     {
