+++ DKBasicEngine 1.0/Core/UI/Button.cs
@@ -2,9 +2,10 @@
 * (C) 2017 David Knieradl 
 */
 
+using DKBasicEngine_1_0.Core.Scripts;
 using System;
 
-namespace DKBasicEngine_1_0
+namespace DKBasicEngine_1_0.Core.UI
 {
     public class Button : TextBlock, IControl
     {
