+++ DKBasicEngine 1.0/Core/UI/Text.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DKBasicEngine_1_0.Core.UI
+{
+    public static class Text
+    {
+        public enum HorizontalAlignment
+        {
+            Left,
+            Center,
+            Right
+        };
+        public enum VerticalAlignment
+        {
+            Top,
+            Center,
+            Bottom
+        };
+
+        public enum InputType
+        {
+            All,
+            AlphaNumerical,
+            Alpha,
+            Numerical
+        };
+    }
+}
