+++ DKBasicEngine 1.0/DKBasicEngine 1.0.csproj
@@ -40,16 +40,17 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Build-in stuff\Scripts\ButtonScript.cs" />
-    <Compile Include="Build-in stuff\Scripts\TextBoxScript.cs" />
-    <Compile Include="Build-in stuff\Scripts\TextControlScript.cs" />
+    <Compile Include="Core\Components\Material.cs" />
+    <Compile Include="Core\Scripts\ButtonScript.cs" />
+    <Compile Include="Core\Scripts\TextBoxScript.cs" />
+    <Compile Include="Core\Scripts\TextControlScript.cs" />
     <Compile Include="Core\Components\Animator.cs" />
     <Compile Include="Core\Components\Camera.cs" />
     <Compile Include="Core\Components\Collider.cs" />
     <Compile Include="Core\Components\AnimationNode.cs" />
-    <Compile Include="Core\Components\Unreleased\MaterialExt.cs" />
+    <Compile Include="Core\Components\Unused\Material_LowMemory.cs" />
     <Compile Include="Core\Components\Transform.cs" />
-    <Compile Include="Core\Components\Material.cs" />
+    <Compile Include="Core\Components\Unused\Material_HighMemory.cs" />
     <Compile Include="Core\Components\Vector3.cs" />
     <Compile Include="Core\Interfaces\I3Dimensional.cs" />
     <Compile Include="Core\Interfaces\IAnimated.cs" />
@@ -63,19 +64,20 @@
     <Compile Include="Core\Interfaces\Random\OwO\K\ICloseTheInternet.cs" />
     <Compile Include="Core\Interfaces\Random\Q-Q\K\ISlepice.cs" />
     <Compile Include="Core\Interfaces\Random\UwU\K\IDunno.cs" />
-    <Compile Include="Core\SystemExt\Button.cs" />
-    <Compile Include="Core\SystemExt\xRectangle.cs" />
-    <Compile Include="Core\SystemExt\Letter.cs" />
+    <Compile Include="Core\UI\Text.cs" />
+    <Compile Include="Core\UI\Button.cs" />
+    <Compile Include="Core\UI\xRectangle.cs" />
+    <Compile Include="Core\UI\Letter.cs" />
     <Compile Include="Core\SystemExt\Extensions.cs" />
-    <Compile Include="Core\SystemExt\TextBlock.cs" />
-    <Compile Include="Core\SystemExt\TextBox.cs" />
+    <Compile Include="Core\UI\TextBlock.cs" />
+    <Compile Include="Core\UI\TextBox.cs" />
     <Compile Include="Core\SystemExt\WindowControl.cs" />
-    <Compile Include="Data\Database.cs" />
-    <Compile Include="Data\GameObject.cs" />
-    <Compile Include="Data\Predefined\Border.cs" />
-    <Compile Include="Data\Predefined\SplashScreen.cs" />
-    <Compile Include="Data\Scene.cs" />
-    <Compile Include="Data\Script.cs" />
+    <Compile Include="Core\Database.cs" />
+    <Compile Include="Core\GameObject.cs" />
+    <Compile Include="Data\Border.cs" />
+    <Compile Include="Data\SplashScreen.cs" />
+    <Compile Include="Core\Scene.cs" />
+    <Compile Include="Core\Script.cs" />
     <Compile Include="Engine.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
@@ -89,6 +91,7 @@
     <Content Include="Build-in stuff\Images\DKEngine_splash1.gif" />
     <Content Include="Build-in stuff\Images\DKEngine_splash2.gif" />
     <Content Include="Build-in stuff\Images\DKEngine_splash3.gif" />
+    <None Include="Core\Components\Unused\MaterialBackups" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Build-in stuff\FontFile.MEX" />
@@ -96,6 +99,7 @@
   <ItemGroup>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
+      <SubType>Designer</SubType>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
