+++ DKBasicEngine 1.0/Properties/Resources.resx
@@ -119,18 +119,18 @@
   </resheader>
   <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
   <data name="border" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\build-in stuff\border.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+    <value>..\build-in stuff\images\border.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="DKEngine_splash1" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\build-in stuff\images\dkengine_splash1.gif;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="DKEngine_splash2" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\build-in stuff\images\dkengine_splash2.gif;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
+  </data>
+  <data name="DKEngine_splash3" type="System.Resources.ResXFileRef, System.Windows.Forms">
+    <value>..\build-in stuff\images\dkengine_splash3.gif;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
   <data name="FontFile" type="System.Resources.ResXFileRef, System.Windows.Forms">
     <value>..\build-in stuff\fontfile.mex;System.Byte[], mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
-  </data>
-  <data name="DKEngine_splash1" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Build-in stuff\DKEngine_splash1.gif;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="DKEngine_splash2" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Build-in stuff\DKEngine_splash2.gif;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="DKEngine_splash3" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>..\Build-in stuff\DKEngine_splash3.gif;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
   </data>
 </root>
